+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/testing/SchemalessKafkaJsonTypeTests.scala
@@ -7,7 +7,7 @@ import com.dimafeng.testcontainers.{
   MultipleContainers,
   SchemaRegistryContainer
 }
-import com.typesafe.config.Config
+import com.typesafe.config.{Config, ConfigValueFactory}
 import com.typesafe.config.ConfigValueFactory.fromMap
 import com.typesafe.scalalogging.StrictLogging
 import io.circe.{Decoder, Json, JsonObject}
@@ -91,6 +91,12 @@ class SchemalessKafkaJsonTypeTests
     }
   }
 
+  "The endpoint for adhoc test parameters should" - {
+    "return test parameters" in {
+      shouldProperlyGetTestParameters()
+    }
+  }
+
   "The endpoint for process validation should" - {
     "validate scenario properly with Json data and return proper typing" in {
       val typedJsonDataSamples = List("{}", "[]", "null").map(getScenarioWithDataSample)
@@ -145,6 +151,10 @@ class SchemalessKafkaJsonTypeTests
       "scenarioTypes.streaming.modelConfig.components.kafka.config.kafkaProperties",
       fromMap(Map("bootstrap.servers" -> "localhost:8070", "schema.registry.url" -> "http://localhost:8069").asJava)
     )
+    .withValue(
+      "scenarioTypes.streaming.modelConfig.components.kafka.config.useDataSampleParamForSchemalessJsonTopicBasedKafkaSource",
+      ConfigValueFactory.fromAnyRef(true)
+    )
 
   lazy val defaultKafkaConfig: KafkaConfig = KafkaConfig(
     kafkaProperties = Some(Map("bootstrap.servers" -> kafkaContainer.bootstrapServers)),
@@ -191,9 +201,9 @@ object SchemalessKafkaJsonTypeTests {
 
   private[SchemalessKafkaJsonTypeTests] trait WithSchemalessAdHocTestParameters extends WithAdHocTestParameters {
 
-    protected def exampleScenarioSourceId: String = "start"
+    override protected def exampleScenarioSourceId: String = "start"
 
-    protected def exampleScenario: CanonicalProcess = {
+    override protected def exampleScenario: CanonicalProcess = {
       ScenarioBuilder
         .streaming("without-schema")
         .parallelism(1)
@@ -201,7 +211,8 @@ object SchemalessKafkaJsonTypeTests {
           exampleScenarioSourceId,
           "kafka",
           "Topic"        -> s"'$sourceTopicName'".spel,
-          "Content type" -> "'JSON'".spel
+          "Content type" -> "'JSON'".spel,
+          "Data sample"  -> Expression.json("{\"name\": \"Tom\"}")
         )
         .emptySink(
           "end",
@@ -215,18 +226,18 @@ object SchemalessKafkaJsonTypeTests {
         )
     }
 
-    protected def validParameters: TestSourceParameters =
+    override protected def validParameters: TestSourceParameters =
       TestSourceParameters(exampleScenarioSourceId, Map(inputParamName -> Expression.json(validJson)))
 
-    protected def invalidParameters: TestSourceParameters =
+    override protected def invalidParameters: TestSourceParameters =
       TestSourceParameters(exampleScenarioSourceId, Map(inputParamName -> Expression.json(invalidJson)))
 
-    protected def parametersProvidedForDryRun: String = AdhocTestParametersRequest(
+    override protected def parametersProvidedForDryRun: String = AdhocTestParametersRequest(
       sourceParameters = validParameters,
       scenarioGraph = toScenarioGraph(exampleScenario)
     ).asJson.toString()
 
-    protected def expectedValidationErrorsOnInvalidParametersJson: String =
+    override protected def expectedValidationErrorsOnInvalidParametersJson: String =
       s"""
          |[
          |  {
@@ -238,6 +249,43 @@ object SchemalessKafkaJsonTypeTests {
          |    "details": null
          |  }
          |]""".stripMargin
+
+    override protected def expectedTestParametersJson: String = {
+      s"""
+         |[
+         |  {
+         |    "sourceId": "$exampleScenarioSourceId",
+         |    "parameters": [
+         |      {
+         |        "name": "Input",
+         |        "typ": {
+         |          "display": "Json",
+         |          "type": "Unknown",
+         |          "refClazzName": "java.lang.Object",
+         |          "params": []
+         |        },
+         |        "editors": [
+         |          {
+         |            "type": "JsonParameterEditor"
+         |          }
+         |        ],
+         |        "defaultValue": {
+         |          "language": "json",
+         |          "expression": "{\\n  \\"name\\" : \\"Tom\\"\\n}"
+         |        },
+         |        "additionalVariables": {},
+         |        "variablesToHide": [],
+         |        "branchParam": false,
+         |        "hintText": null,
+         |        "label": "Value",
+         |        "requiredParam": true,
+         |        "category": "Standard"
+         |      }
+         |    ]
+         |  }
+         |]
+         |""".stripMargin
+    }
 
     private val validJson = """|[
                                |  {